add usage example draw-ellipse
