--- public/usage-examples/graphics/draw_circle_on_window/draw_circle_on_window-1-random-color-circle.txt
@@ -1,3 +1,0 @@
-#### Draw Circle On Window
-
-The following code shows examples of using [Draw Circle On Window](/api/graphics#draw_circle_on_window-1) to show how to draw multiple random color and random size circles on window.
