+++ public/usage-examples/graphics/draw_ellipse/draw_ellipse-1-radiating-wave.txt
@@ -1,0 +1,3 @@
+#### Draw Ellipse
+
+The following code shows examples of using [Draw Ellipse](/api/graphics#draw-ellipse-1) to show how to use the function draw ellipse to draw radiating wave.
